Combine properties monoidally
Besides a nicer syntax for setting multiple properties, basing the
property bundle types upon `Data.Monoid.Last` offers a way to leave
properties unspecified. That, in turn, makes `modifyFlex`, a combinator
which preserves unspecified properties, possible.

The point of `defaultFlexProps` is making it possible to reimplement
`setFlex` while changing the interface as little as possible. The
`FlexProps` subclass of `ToStyle` it belongs to exists primarily to
avoid having to supply a meaningless `defaultFlexProps` implementation
for `Clay.Stylesheet.Rule`.